Smart Article Categorizer/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.1
scikit-learn==1.4.1
numpy==1.26.4
pandas==2.2.1
transformers==4.37.2
sentence-transformers==2.3.1
gensim==4.3.2
openai==1.12.0
matplotlib==3.8.3
seaborn==0.13.2
plotly==5.18.0
python-multipart==0.0.9
torch==2.2.0
tqdm==4.66.2
umap-learn==0.5.5 